# future/selectable: lazy allocate wait chan