Before:
BenchmarkSelectable-4   	 1000000	      1085 ns/op
BenchmarkSelectable 	 2000000	       983 ns/op
After:
BenchmarkSelectable-4   	 2000000	       991 ns/op
BenchmarkSelectable 	 2000000	       885 ns/op